Enhance looping controllers adding shift buttons to emulate default behaviour.
This commit adds the use of SHIFT in loops as desccribed in the users manual:
Press SHIFT + pad 1 to create the start of the loop.
Press SHIFT + pad 2 to mark the end of the loop or to exit the loop.
Press SHIFT + pad 3 to divide the loop length in half.
Press SHIFT + pad 4 to double the length of the loop.